feat(tech-sheets): Automatitzar l'actualització del personal tècnic
- Elimina el botó d'actualització manual des de les assignacions a la fitxa de bolo.
- Implementa una actualització automàtica del personal tècnic a la fitxa de bolo quan canvien les assignacions de l'esdeveniment.
- Millora la lògica per a esdeveniments de diversos dies: ara, a les notes del personal amb assignacions mixtes, s'especifiquen els dies concrets en què estan confirmats.
- La secció de personal tècnic ara es sincronitza automàticament amb les dades de les assignacions, de manera similar a com ho fa la informació general de l'esdeveniment.
